# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v2/auth/device/{deviceId}": {
            "delete": {
                "description": "吊销指定设备的访问权限",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "吊销设备",
                "parameters": [
                    {
                        "type": "string",
                        "description": "设备ID",
                        "name": "deviceId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v2/auth/refresh": {
            "post": {
                "description": "使用刷新令牌获取新的访问令牌",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "刷新访问令牌",
                "parameters": [
                    {
                        "description": "刷新令牌请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.RefreshTokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.TokenResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v2/auth/register": {
            "post": {
                "description": "设备注册，获取访问令牌和证书",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "注册设备",
                "parameters": [
                    {
                        "description": "设备注册信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.DeviceRegistration"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v2/auth/token": {
            "post": {
                "description": "根据租户ID和设备密钥获取访问令牌(仅用于测试)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "获取访问令牌",
                "parameters": [
                    {
                        "description": "令牌请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.TokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.TokenResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v2/auth/validate/{deviceId}": {
            "get": {
                "description": "验证设备是否已注册且有效",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "验证设备",
                "parameters": [
                    {
                        "type": "string",
                        "description": "设备ID",
                        "name": "deviceId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ValidateDeviceResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v2/tenants/{tenantId}/audit-logs": {
            "get": {
                "description": "查询指定租户的审计日志记录",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "审计日志"
                ],
                "summary": "查询审计日志",
                "parameters": [
                    {
                        "type": "string",
                        "description": "租户ID",
                        "name": "tenantId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 50,
                        "description": "每页数量",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "偏移量",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "操作者ID",
                        "name": "actor_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "操作类型",
                        "name": "action",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "资源类型",
                        "name": "resource_type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "error message"
                }
            }
        },
        "api.RefreshTokenRequest": {
            "type": "object",
            "properties": {
                "refresh_token": {
                    "type": "string",
                    "example": "xxx"
                }
            }
        },
...
        "api.TokenRequest": {
            "type": "object",
            "properties": {
                "device_key": {
                    "type": "string",
                    "example": "device-key-xxx"
                },
                "tenant_id": {
                    "type": "string",
                    "example": "tenant001"
                }
            }
        },
        "api.TokenResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                },
                "expires_at": {
                    "type": "integer",
                    "example": 1700000000000
                },
                "refresh_token": {
                    "type": "string",
                    "example": "xxx"
                },
                "token_type": {
                    "type": "string",
                    "example": "Bearer"
                }
            }
        },
...
        "api.ValidateDeviceResponse": {
            "type": "object",
            "properties": {
                "device_id": {
                    "type": "string",
                    "example": "device001"
                },
                "valid": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "models.DeviceInfo": {
            "type": "object",
            "properties": {
                "android_version": {
                    "type": "string"
                },
                "app_version": {
                    "type": "string"
                },
                "battery_charging": {
                    "type": "boolean"
                },
                "battery_level": {
                    "type": "integer"
                },
                "brand": {
                    "type": "string"
                },
                "free_memory": {
                    "type": "integer"
                },
                "free_storage": {
                    "type": "integer"
                },
                "imei": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "screen_brightness": {
                    "type": "integer"
                },
                "screen_on": {
                    "type": "boolean"
                },
                "sdk_version": {
                    "type": "integer"
                },
                "serial_number": {
                    "type": "string"
                },
                "total_memory": {
                    "type": "integer"
                },
                "total_storage": {
                    "type": "integer"
                },
                "volume": {
                    "type": "integer"
                },
                "wifi_mac": {
                    "type": "string"
                }
            }
        },
        "models.DeviceRegistration": {
            "type": "object",
            "properties": {
                "csr_pem": {
                    "type": "string"
                },
                "device_info": {
                    "$ref": "#/definitions/models.DeviceInfo"
                },
                "device_key": {
                    "type": "string"
                },
                "integrity_token": {
                    "type": "string"
                },
                "nonce": {
                    "type": "string"
                },
                "tenant_id": {
                    "type": "string"
                }
            }
        },
...
    }
}